PWM setup & diagnostics (#5)
This PR adds a library for controlling the am335x ehrpwm module. This is
a work-in-progress, having demonstrated how to configure the basic
settings and see a functional PWM output. This PR also demonstrates a
failure to make PWM interrupt generation work for the PRU.

Will post a forum question using this content.

See comments in lib/pwm/am335x/pwm-am335x.c.